+++ service/src/constants.rs
@@ -6,4 +6,5 @@ use once_cell::sync::Lazy;
 use std::path::Path;
 
 pub const OBJ_PATH_STR: &str = "/com/musikid/fancy";
+pub const BUS_NAME_STR: &str = "com.musikid.fancy";
 pub static ROOT_CONFIG_PATH: Lazy<&Path> = Lazy::new(|| Path::new("/etc/fancy"));